Completed basic life cycle for a single job
With the server running, you can add a job and if it returns
true, the action will execute and removed from the queue

+++ sensors/scale.go
@@ -20,7 +20,7 @@ func (Scale) setupScale() *Scale {
 }
 
 // GetWeight gets the current weight value from teh scale sensor
-func (*Scale) GetWeight() int {
+func (*Scale) GetWeight() float64 {
 	// TODO: get weight from sensor
 	return 200
 }
